+++ backend/fit-journey-api/src/main/java/hu/hm/fitjourneyapi/repository/UserRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     List<User> findUsersByName(String name);
     Optional<User> findUserByEmail(String email);
 
-    Optional<User> findUserById(int id);
+    Optional<User> findUserById(long id);
 
-    User getUserById(int id);
+    User getUserById(long id);
 }
