+++ backend/fit-journey-api/src/main/java/hu/hm/fitjourneyapi/repository/fitness/ExerciseRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ExerciseRepository extends JpaRepository<Exercise, Integer> {
+public interface ExerciseRepository extends JpaRepository<Exercise, Long> {
     Optional<Object> findByName(String name);
 }
