+++ backend/fit-journey-api/src/main/java/hu/hm/fitjourneyapi/repository/fitness/FlexibilitySetRepository.java
@@ -3,5 +3,5 @@ package hu.hm.fitjourneyapi.repository.fitness;
 import hu.hm.fitjourneyapi.model.fitness.FlexibilitySet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FlexibilitySetRepository extends JpaRepository<FlexibilitySet, Integer> {
+public interface FlexibilitySetRepository extends JpaRepository<FlexibilitySet, Long> {
 }
