+++ backend/fit-journey-api/src/main/java/hu/hm/fitjourneyapi/repository/fitness/StrengthSetRepository.java
@@ -3,6 +3,6 @@ package hu.hm.fitjourneyapi.repository.fitness;
 import hu.hm.fitjourneyapi.model.fitness.StrengthSet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StrengthSetRepository extends JpaRepository<StrengthSet, Integer> {
+public interface StrengthSetRepository extends JpaRepository<StrengthSet, Long> {
     
 }
